Add optimus-manager.conf, sddm.conf, new packages
Some of the keys on my keyboard were getting warm, even when not doing
anything GPU intensive. This happens because when the system is in
hybrid mode, the GPU keeps consuming power, even when no applications
are running on it. This means less battery life and more heat.

After installing optimus-manager and configuring it to use the Nouveau
driver when in Intel integrated graphics mode (see:
https://github.com/Askannz/optimus-manager/wiki/A-guide--to-power-management-options):
the laptop produces a noticeable amount of less heat, which is quite nice. To switch
between graphics modes, type `optimus-manager --switch [intel|hybrid|nvidia]` in the
terminal.