Deprecate instructions in sound.txt
It turns out that I was a bit of an idiot and that you don't actually
NEED to add snd_hda_intel.dmic_detect=0 to the kernel parameters in
order to have sound. All I needed to do was to unmute some channels
which were muted in alsamixer -_-.

Anyway, to get sound, unmute these channels in alsamixer (Playback tab):
Master
Headphones
Speaker (I set this to 40% as the internal speakers are louder than
I expected, but YMMV)
PGA1.0 1 Master
Also, enable Auto-Mute mode so that the internal speakers automatically
mute themselves when you plug in headphones or external speakers into
the headphone port (unless you don't want that for whatever reason)

To set microphone output to a normal level, in the capture tab:
Capture (100%)
PGA2.0 Master (30% or whereabouts)

Otherwise, you will hear VERY obvious background noise.