Change XDG_CURRENT_DESKTOP comment in start-dwm.sh
Changed the XDG_CURRENT_DESKTOP comment in start-dwm.sh as this was the
easiest way of doing it.

As an alternative, you can setup the appearance through qt5ct, but from
my testing that required setting two environment variables:
QT_STYLE_OVERRIDE and QT_QPA_PLATFORMTHEME. Setting just
QT_QPA_PLATFORMTHEME did 99% of the work, but it left a weird looking
button (gray with gradient) in the lower right corner of Dolphin.
Additionally setting QT_STYLE_OVERRIDE fixed that but that would mean
that I would have to change that export in start-dwm if I use another
theme (so if I decide to use some other theme than Breeze, I would have
to change the export QT_STYLE_OVERRIDE=... line)

So I left it as it is - it works just fine.