Add available-updates script
Added an available-updates script, which checks the number of pending
updates and shows them as a notification on the screen.

Requires pacman-contrib (for `checkupdates`) and a notification server
(such as dunst or KDE's built-in one).

For now, dunst doesn't show any icons or action buttons when running
under dwm, I will mess around with its config file later.